修复错误，顺便开源一个正在做的远程预览程序

+++ 安卓/src/main/java/间/安卓/工具/转换.java
@@ -32,8 +32,8 @@ public class 转换 {
         return 转换大小(TypedValue.COMPLEX_UNIT_IN, $值);
     }
 
-    public static Float 数字(String $值) {
-        return new Float($值);
+    public static int 数字(String $值) {
+        return new Integer($值);
     }
     
     private static Float 转换大小(int $类型,int $值) {
